// frontend/src/app/api/models/match-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MatchStickerDto } from '../models/match-sticker-dto';
export interface MatchResponse {
  exchangeableCount?: number;
  itemsOffered?: Array<MatchStickerDto>;
  itemsRequested?: Array<MatchStickerDto>;
  userId?: number;
}